+++ R/survival-coxph-tidiers.R
@@ -61,43 +61,31 @@
 #' @family survival tidiers
 tidy.coxph <- function(x, exponentiate = FALSE, conf.int = FALSE,
                        conf.level = .95, ...) {
-  # backward compatibility (in previous version, conf.int was used instead of conf.level)
-  if (is.numeric(conf.int)) {
-    conf.level <- conf.int
-    conf.int <- TRUE
-  }
-
-  if (conf.int) {
-    s <- summary(x, conf.int = conf.level)
-  } else {
-    s <- summary(x, conf.int = FALSE)
-  }
+ 
+  s <- summary(x)
   co <- stats::coef(s)
 
   if (!is.null(x$frail)) {
     nn <- c("estimate", "std.error", "statistic", "p.value")
-  } else if (s$used.robust) {
+  } else if (isTRUE(s$used.robust)) {
     nn <- c("estimate", "std.error", "robust.se", "statistic", "p.value")
   } else {
     nn <- c("estimate", "std.error", "statistic", "p.value")
   }
 
-  if (is.null(x$frail)) {
+  if (is.null(x$frail) && is.null(x$penalty)) {
     ret <- as_tidy_tibble(co[, -2, drop = FALSE], new_names = nn)
   } else {
     ret <- as_tidy_tibble(co[, -c(3, 5), drop = FALSE], new_names = nn)
   }
 
-  if (exponentiate) {
-    ret$estimate <- exp(ret$estimate)
+  if (conf.int) {
+    ci <- broom_confint_terms(x, level = conf.level)
+    ret <- dplyr::left_join(ret, ci, by = "term")
   }
-  if (!is.null(s$conf.int)) {
-    CI <- as.matrix(unrowname(s$conf.int[, 3:4, drop = FALSE]))
-    colnames(CI) <- c("conf.low", "conf.high")
-    if (!exponentiate) {
-      CI <- log(CI)
-    }
-    ret <- cbind(ret, CI)
+
+  if (exponentiate) {
+    ret <- exponentiate(ret)
   }
 
   as_tibble(ret)