+++ NEWS.md
@@ -1,11 +1,65 @@
-# broom 0.7.2.9000
+# broom 0.7.5.9000
 
-To be released as 0.7.3.
+To be released as broom 0.7.6.
+
+* Fixed bug in `augment` tidiers resulting in `.fitted` and `.se.fit` array columns.
+* Fixed bug that made column `y` non-numeric after `tidy_xyz` (`#973` by `@jiho`)
+* Add tidiers for `MASS:glm.nb` (`#998` by `@joshyam-k`)
+* Fixed bug in `tidy.fixest` that sometimes prevented arguments like `se` from being used (`#1001` by `@karldw`)
+
+# broom 0.7.5
+
+* Fixed bug in the `nnet::multinom` tidier in the case that the response
+variable has only two levels (`#993` by `@vincentarelbundock` and `@hughjonesd`)
+* Various bug fixes and improvements to documentation
+
+# broom 0.7.4
+
+broom 0.7.4 introduces tidier support for a number of new model objects and 
+improves functionality of many existing tidiers!
+
+#### New Tidiers
+
+* Add tidiers for `Rchoice` objects (`#961` by `@vincentarelbundock` and `@Nateme16`)
+* Add tidiers for objects produced by `car::leveneTest` (`#968` by `@vincentarelbundock` and `@mkirzon`)
+* Add tidiers for objects produced by `cmprsk::crr` (`#971` and `#552` by `@vincentarelbundock` and `@margarethannum`)
+* Add an `augment()` method for `gam` objects (`#975` and `#645` by `@vincentarelbundock`)
+* Add tidiers for `vars` objects (`#979` and `#161` by `@vincentarelbundock` and `@Diego-MX`)
+
+This release also restores tidiers for `felm` objects from the `lfe` package, which was recently unarchived from CRAN.
+
+#### Improvements to existing tidiers
+
+* `tidy.emmGrid` can now return `std.error` and `conf.*` columns at the same time. (`#962` by `@vincentarelbundock` and `@jmbarbone`)
+* `tidy.garch` can now produce confidence intervals (`#964` by `@vincentarelbundock` and `@IndrajeetPatil`)
+* `tidy.coxph` can now report confidence intervals on models utilizing penalized/clustering terms (`#966` by `@vincentarelbundock` and `@matthieu-faron`)
+* `augment.lm` now works when some regression weights are equal to zero (`#965` by `@vincentarelbundock` and `@vnijs`)
+* `tidy.coxph` can now handle models utilizing penalized/clustering terms (`#966` and `#969` by `@vincentarelbundock`, `@matthieu-faron`, and `@KZARCA`)
+* Fix bug in `tidy.speedglm` on R 4.0.0+ (`#974` by `@uqzwang`)
+* tidy.multinom works with matrix response (`#977` and `#666` by `@vincentarelbundock` and `@atyre2`)
+* Various bug fixes and improvements to documentation and errors.
+
+# broom 0.7.3
+
+In broom `0.7.0`, we introduced an error for model objects that subclassed
+`lm` and relied on `tidy.lm()`, or similarly for `tidy.glm()`. Tidiers for
+these objects were supported unintentionally, and we worried that tidiers for
+these objects would silently report innaccurate results.
+
+In hindsight, this change was unnecessarily abrupt. We've decided to roll back 
+this change, instead providing the following warning before allowing such 
+objects to fall back to the `lm`/`glm` tidier methods:
+
+> Tidiers for objects of class {subclass} are not maintained by the broom team, and are only supported through the {dispatched_method} tidier method. Please be cautious in interpreting and reporting broom output."
+
+In addition,
+
+* Restores tidiers for `summary.lm` objects (`#953` by `@grantmcdermott`)
+* Deprecate tidiers for the `lfe` package, which was archived from CRAN.
+* Various bug fixes and improvements to documentation and errors.
 
 # broom 0.7.2
 
-* In broom `0.7.0`, we introduced an error for objects that subclassed
-`lm` and relied on `tidy.lm()`. We've transitioned this error to a warning.
 * Various bug fixes and improvements to documentation and errors.
 
 # broom 0.7.1